% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/su_palette.R
\docType{data}
\name{su_palettes}
\alias{callettes}
\title{callettes}
\format{
  An object of class \code{list} of length 7.
}
\usage{
  callettes
}
\description{
  Color palettes by Calle: `callettes`
  }
\details{
  (code based on https://github.com/jrcalabrese/joycon)
  
  Options: \code{rainbow}, \code{friendly},
  \code{red}, \code{orange}, \code{green},
  \code{blue}, \code{indigo}, \code{violet}
}
\keyword{datasets}
